+++ CHANGELOG.md
@@ -8,7 +8,7 @@ We used to have incremental versioning before `0.1.0`.
 
 ### Features
 
-- `WPS431` now allow customize whitelist via  `nested-classes-whitelist` setting
+- `WPS431` now allow customize whitelist via `nested-classes-whitelist` setting
 
 ## 0.12.5